LayoutTests/platform/mac-bigsur/imported/w3c/web-platform-tests/upgrade-insecure-requests/gen/iframe-blank-inherit.meta/upgrade/worker-import-data.https-expected.txt

Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/worker.py?redirection=downgrade&action=purge&key=GENERATED_KEY&path=/mixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/worker.py?redirection=no-redirect&action=purge&key=GENERATED_KEY&path=/mixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/worker.py?redirection=downgrade&action=purge&key=GENERATED_KEY&path=/mixed-content

FAIL Upgrade-Insecure-Requests: Expects allowed for worker-import-data to cross-http-downgrade origin and downgrade redirection from https context. assert_equals: The resource request should be 'allowed'. expected "allowed" but got "blocked"
FAIL Upgrade-Insecure-Requests: Expects allowed for worker-import-data to cross-http-downgrade origin and no-redirect redirection from https context. assert_equals: The resource request should be 'allowed'. expected "allowed" but got "blocked"
FAIL Upgrade-Insecure-Requests: Expects allowed for worker-import-data to cross-https origin and downgrade redirection from https context. assert_equals: The resource request should be 'allowed'. expected "allowed" but got "blocked"
PASS Upgrade-Insecure-Requests: Expects allowed for worker-import-data to same-http-downgrade origin and downgrade redirection from https context.
PASS Upgrade-Insecure-Requests: Expects allowed for worker-import-data to same-http-downgrade origin and no-redirect redirection from https context.
PASS Upgrade-Insecure-Requests: Expects allowed for worker-import-data to same-https origin and downgrade redirection from https context.
